memo_server/requirements.txt
flask (>= 3.0.0)
requests (>= 2.0.0)
redis (>= 5.0.0)
pymongo (>= 2.4.0)
mysql-connector-python (>= 8.0.0)